split structs into seperate files
